revert context test change because it breaks build
