+++ codecov.yml
@@ -19,8 +19,7 @@ coverage:
         if_ci_failed: ignore   # require the CI to pass before setting the status
     patch:
       default:
-        target: 0% # Todo (David.porter) to immediately revert this after landing a refactor
-                               # specify the target coverage for each commit status
+        target: 80%            # specify the target coverage for each commit status
                                #   option: "auto" (compare against parent commit or pull request base)
                                #   option: "X%" a static target percentage to hit
         threshold: 0%          # allow the coverage drop by x% before marking as failure
