+++ service/history/shard/contextTest.go
@@ -50,91 +50,6 @@ import (
 	"github.com/uber/cadence/service/history/resource"
 )
 
-// TestContext is a test implementation for shard Context interface
-type TestContext struct {
-	*contextImpl
-
-	Resource                        *resource.Test
-	MockEventsCache                 *events.MockCache
-	MockAddingPendingFailoverMarker func(*types.FailoverMarkerAttributes) error
-}
-
-var _ Context = (*TestContext)(nil)
-
-// NewTestContext create a new shardContext for test
-func NewTestContext(
-	t *testing.T,
-	ctrl *gomock.Controller,
-	shardInfo *persistence.ShardInfo,
-	config *config.Config,
-) *TestContext {
-	resource := resource.NewTest(t, ctrl, metrics.History)
-	eventsCache := events.NewMockCache(ctrl)
-	if shardInfo.TransferProcessingQueueStates == nil {
-		shardInfo.TransferProcessingQueueStates = &types.ProcessingQueueStates{
-			StatesByCluster: make(map[string][]*types.ProcessingQueueState),
-		}
-	}
-	if shardInfo.TimerProcessingQueueStates == nil {
-		shardInfo.TimerProcessingQueueStates = &types.ProcessingQueueStates{
-			StatesByCluster: make(map[string][]*types.ProcessingQueueState),
-		}
-	}
-	if shardInfo.CrossClusterProcessingQueueStates == nil {
-		shardInfo.CrossClusterProcessingQueueStates = &types.ProcessingQueueStates{
-			StatesByCluster: make(map[string][]*types.ProcessingQueueState),
-		}
-	}
-	shard := &contextImpl{
-		Resource:                  resource,
-		shardID:                   shardInfo.ShardID,
-		rangeID:                   shardInfo.RangeID,
-		shardInfo:                 shardInfo,
-		executionManager:          resource.ExecutionMgr,
-		config:                    config,
-		logger:                    resource.GetLogger(),
-		throttledLogger:           resource.GetThrottledLogger(),
-		transferSequenceNumber:    1,
-		transferMaxReadLevel:      0,
-		maxTransferSequenceNumber: 100000,
-		timerMaxReadLevelMap:      make(map[string]time.Time),
-		remoteClusterCurrentTime:  make(map[string]time.Time),
-		eventsCache:               eventsCache,
-	}
-	return &TestContext{
-		contextImpl:     shard,
-		Resource:        resource,
-		MockEventsCache: eventsCache,
-	}
-}
-
-// ShardInfo is a test hook for getting shard info
-func (s *TestContext) ShardInfo() *persistence.ShardInfo {
-	return s.shardInfo
-}
-
-// SetEventsCache is a test hook for setting events cache
-func (s *TestContext) SetEventsCache(
-	eventsCache events.Cache,
-) {
-	s.eventsCache = eventsCache
-	s.MockEventsCache = nil
-}
-
-// Finish checks whether expectations are met
-func (s *TestContext) Finish(
-	t mock.TestingT,
-) {
-	s.Resource.Finish(t)
-}
-
-func (s *TestContext) AddingPendingFailoverMarker(marker *types.FailoverMarkerAttributes) error {
-	if s.MockAddingPendingFailoverMarker != nil {
-		return s.MockAddingPendingFailoverMarker(marker)
-	}
-	return s.contextImpl.AddingPendingFailoverMarker(marker)
-}
-
 const (
 	testShardID                   = 123
 	testRangeID                   = 1
